removed MenuPicker platform check and added MenuPicker platform specific files
