+++ apps/fluent-tester/src/FluentTester/TestComponents/Common/MenuPicker.macos.tsx
@@ -1,0 +1,1 @@
+export * from './MenuPicker.desktop';
