+++ apps/fluent-tester/src/FluentTester/TestComponents/Common/MenuPicker.win32.tsx
@@ -1,0 +1,1 @@
+export * from './MenuPicker.desktop';
